solution/backend/requirements.txt
cryptography>=42.0.2
pyopenssl>=24,<25
Werkzeug
boto3
django-debug-toolbar
django>=5.1
requests
djangorestframework
psycopg2-binary
django-solo
python-dateutil
django-cors-headers
drf-spectacular
django-model-utils
django-jsonform
django-csp
httpx
ruff
pytest
pytest-django
pre-commit
defusedxml
mozilla-django-oidc
boto3
django-storages
pytest-cov
djangorestframework_simplejwt
olefile
sqlparse>=0.5.0 # not directly required, pinned by Snyk to avoid a vulnerability
natsort==8.4.0
